# Conflicts:
#	Assets/Scripts/Dialogue/Dialogue.cs.meta
#	Assets/Scripts/InventorySystem/Item Data/Coin.asset.meta
#	Assets/Scripts/InventorySystem/Item Data/GoldApple.asset.meta
#	Assets/Scripts/InventorySystem/Item Data/Sword.asset.meta
#	Assets/Scripts/InventorySystem/ToolTip/ToolTipSystem.cs.meta
#	Assets/Scripts/InventorySystem/ToolTip/ToolTipTrigger.cs.meta

+++ Assets/Scripts/ObjectHealth.cs
@@ -6,7 +6,7 @@ public class ObjectHealth : MonoBehaviour
 {
     public bool damagable = true; //if the object will die or not
     public bool giveUpwardForce = true; // can the player bounce on it
-    [SerializeField] private int health; //how many hits before dying
+    [SerializeField] private float health; //how many hits before dying
     [SerializeField] private GameObject BloodEffect; //animation for after dying
 
     void Start()
@@ -22,7 +22,7 @@ public class ObjectHealth : MonoBehaviour
         }
     }
 
-    public void DealDamage(int damage)
+    public void DealDamage(float damage)
     {
         Destroy(Instantiate(BloodEffect, transform.position, Quaternion.identity), 1);
         health -= damage;
